block clean "MOICAT.RINDX" table at backend

+++ www/include/Query.class.php
@@ -864,12 +864,11 @@ class Query {
 				$log->info(__METHOD__."：指定刪除 $table 故跳過 $tmp_tbl_name 。");
 				continue;
 			}
-			
+			if ($tmp_tbl_name == "MOICAT.RINDX") {
+				$log->warning(__METHOD__."：無法刪除 MOICAT.RINDX 跳過！");
+				continue;
+			}
 			$log->info(__METHOD__."：刪除 $tmp_tbl_name 資料 ... ");
-			
-			/*$this->db->parse("
-				DELETE FROM ".$tmp_tbl_name." WHERE ".$key_fields[0]." = :bv_year AND ".$key_fields[1]." = :bv_code AND ".$key_fields[2]." = :bv_number
-			");*/
 			$this->db->parse("
 				DELETE FROM ".$tmp_tbl_name." WHERE ".$key."03 = :bv_year AND ".$key."04_1 = :bv_code AND ".$key."04_2 = :bv_number
 			");
